Migrate sin tests
