+++ src/unittests/floating_point.spec.ts
@@ -928,6 +928,40 @@ g.test('signInterval_f32')
     );
   });
 
+g.test('sinInterval_f32')
+  .paramsSubcasesOnly<ScalarToIntervalCase>(
+    // prettier-ignore
+    [
+      // This test does not include some common cases, i.e. f(x = -π|π) = 0,
+      // because the difference between true x and x as a f32 is sufficiently
+      // large, such that the high slope of f @ x causes the results to be
+      // substantially different, so instead of getting 0 you get a value on the
+      // order of 10^-8 away from it, thus difficult to express in a
+      // human-readable manner.
+      { input: kValue.f32.infinity.negative, expected: kAnyBounds },
+      { input: kValue.f32.negative.min, expected: kAnyBounds },
+      { input: kValue.f32.negative.pi.half, expected: [-1, plusOneULPF32(-1)] },
+      { input: 0, expected: 0 },
+      { input: kValue.f32.positive.pi.half, expected: [minusOneULPF32(1), 1] },
+      { input: kValue.f32.positive.max, expected: kAnyBounds },
+      { input: kValue.f32.infinity.positive, expected: kAnyBounds },
+    ]
+  )
+  .fn(t => {
+    const error = (_: number): number => {
+      return 2 ** -11;
+    };
+
+    t.params.expected = applyError(t.params.expected, error);
+    const expected = FP.f32.toInterval(t.params.expected);
+
+    const got = FP.f32.sinInterval(t.params.input);
+    t.expect(
+      objectEquals(expected, got),
+      `f32.sinInterval(${t.params.input}) returned ${got}. Expected ${expected}`
+    );
+  });
+
 interface VectorToIntervalCase {
   input: number[];
   expected: number | IntervalBounds;
