+++ src/webgpu/shader/execution/expression/call/builtin/sin.spec.ts
@@ -10,10 +10,10 @@ Returns the sine of e. Component-wise when T is a vector.
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../../../gpu_test.js';
 import { TypeF32 } from '../../../../../util/conversion.js';
-import { sinInterval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { fullF32Range, linearRange } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { allInputSources, generateUnaryToF32IntervalCases, run } from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -21,14 +21,14 @@ export const g = makeTestGroup(GPUTest);
 
 export const d = makeCaseCache('sin', {
   f32: () => {
-    return generateUnaryToF32IntervalCases(
+    return FP.f32.generateScalarToIntervalCases(
       [
-        // Well defined accuracy range
+        // Well-defined accuracy range
         ...linearRange(-Math.PI, Math.PI, 1000),
         ...fullF32Range(),
       ],
       'unfiltered',
-      sinInterval
+      FP.f32.sinInterval
     );
   },
 });
